Split filtering tests into different test classes
